Update the README, add an all_in_one_pe role
Updating README for instructions on how to use the control-repo
by putting it in your own git server and then using the r10k
answers during installation of PE.

Removing the manifest for configuring r10k using zack/r10k.

Added a role that can be used to bootstrap an all-in-one PE
installation.