Correct Namespacing After Moving out of puppetlabs/control-repo
This module originated as a profile in the puppetlabs/control-repo.
Turns out its useful enough to be used on its own so we're breaking
it out.

All the reference to profile/git_webhook are now pe_git_webhook

The functions that were their own modules are going to ship in the
lib dir of this module for simplicity.