Allow configuring of the git repo name
Prior to this commit, the git repo project name was hardcoded to
puppet/control-repo because it was only used with the
puppetlabs/control-repo and the instructions there said to use that
name.

After this commit, the default is still puppet/control-repo but
a user can now set the value as a parameter.