Add a check to see if code manager is enabled via NC or hiera
Prior to this commit the determination of whether code manager
or zack/r10k should be configured was based on PE version.

After this commit, since code manager needs to be enabled for all
of this to work, we'll check if code manager is enabled via the
PE console UI or via hiera and then only proceed with configuring
code manager if that is the case.

If you are using 2015.3+ and have not enabled code manager yet then
you will be notified to do so instead of automatically installing
zack/r10k.  Use $force_zack_r10k_webhook if you prefer.