Comment out node_query function use and instead default to hiera
Prior to this commit I was considering querying the NC to see if
code manager was enabled.  However, the requirement of installing
the puppetclassify gem seems like an undue burden for a small check.

After this commit, we'll rely on the hiera value and if a user
prefers to enable code manager via the PE console then they can
provide the force_code_manager_webhook parameter to enable it.
This passes the extra burden to only the users who place the
parameter in the NC as opposed to all users.