Change ssh key location for code manager to match PE docs
Resolving https://github.com/puppetlabs/control-repo/issues/14

Prior to this commit the location of the ssh key that code manager
uses to connect to your git repo did not match what the PE docs
recommend.

After this commit, the location does match and to maintain backwards
compatibility we move the ssh key from the old location if it is
already there.

Also, adding the ability to disable creating and managing the ssh
key.  This is not recommended as the permissions of the key are
important and disabling this means you should be managing the key
elsewhere to make sure the permissions are correct.