Stop trying to move the old ssh key to the new location
Prior to this commit we tried to move the old suggested location
for the ssh key to the new location.  However, the key that code
manager should use is in the node classifier and we don't have an
easy way to introspect that.

After this commit, we'll leave the old key unmanaged but still present.
Previous users should change the key code manager uses to match the
new location but will continue to work if they do not.

Changed the name of the deploy key so that a new deploy key will
be made in the git system of choice and leave the old one in place.