Allow reading the gms_api_token from a file on disk
Prior to this commit, the only way to set the gms_api_token was
via hiera or as a parameter.  Since the puppetlabs/control-repo
uses a role I cannot inject a parameter into this module.

After this commit, you can supply the gms_api_token from a file
on disk.

I'm not sure if anyone really needs this other than me for automating
the testing of this module and the puppetlabs/control-repo using
npwalker/control-repo-vagrant.  But now that we have this ability
the vagrant stack can loy down the gms_api_token file and the
vagrant stack will not need manual intervention to work.