Fix code manager auto configuration hiera default value
Prior to this commit, if you did not have the hiera key for
code manager set then your puppet run would fail because there
was not default set on the hiera lookup.

After this commit, the hiera lookup for the code mangaer auto
configure key will default to false.

Also, fixed the notify to not interpolate variables by single
quoting it.