Fix hiera lookup for code_manager_auto_configure to default to undef
Prior to this commit, the lookup for code_manager_auto_configure
would default to false which caused the subseqent pick statement
to always return false if the value is not set to true in hiera.

After this commit, the hiera lookup will default to undef if not
specified in hiera data and will allow the pick to function
correctly.