Correct dependencies in metadata.json
Prior to this commit, if you used puppet module install to install
this module it would not install pltraining-rbac which is required
for it to work.

After this commit, pltraining-rbac and abrader-gms have been added
to the dependencies.  abrader-gms is not strictly necessary for
the minimum use case but is needed to setup deploy keys and webhooks.