Handle array of hashes in classifier.yaml in 2016.5
Prior to this commit, we assumed that classifier.yaml was simply a
hash that had the server and port we wanted to know about for
connecting to the classifier.

After this commit, we know that classifier.yaml is an array of
hashes in 2016.5.0 and above so we can correctly find the server
and port that we want.