Use code deployer role included with PE
Prior to this commit, we used a curl to the RBAC API to create
a 'deploy environmets' role that our service user can use.
However, in 2016.5 the RBAC API no longer accepts setting the
token override_expiry permission as it has been removed.

After this commit, we simply use the built-in code deployer role
and no longer allow for creation of a role by a different name.
You can still specify a different role name but you'll have to make
that role yourself beforehand.